chore: inline contents of `common/proof.rs`
