+++ common/src/lib.rs
@@ -7,8 +7,6 @@ pub mod barretenberg_structures;
 pub mod crs;
 pub mod merkle;
 
-pub mod proof;
-
 // Re-export acvm
 pub use acvm;
